Move the logic to the otel manager
