+++ internal/pkg/agent/application/coordinator/coordinator_test.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/elastic/elastic-agent-libs/mapstr"
 
-	otelcomponentmanager "github.com/elastic/elastic-agent/internal/pkg/otel/componentmanager"
 	otelmanager "github.com/elastic/elastic-agent/internal/pkg/otel/manager"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/status"
@@ -1079,9 +1078,8 @@ func createCoordinator(t testing.TB, ctx context.Context, opts ...CoordinatorOpt
 	cfg.Port = 0
 	rm, err := runtime.NewManager(l, l, ai, apmtest.DiscardTracer, monitoringMgr, cfg)
 	require.NoError(t, err)
-	otelMgr, err := otelmanager.NewOTelManager(l, logp.InfoLevel, l, otelmanager.EmbeddedExecutionMode)
+	otelMgr, err := otelmanager.NewOTelManager(l, logp.InfoLevel, l, otelmanager.EmbeddedExecutionMode, ai, monitoringMgr.ComponentMonitoringConfig)
 	require.NoError(t, err)
-	otelComponentMgr := otelcomponentmanager.NewOtelComponentManager(l, otelMgr, ai, monitoringMgr.ComponentMonitoringConfig)
 	caps, err := capabilities.LoadFile(paths.AgentCapabilitiesPath(), l)
 	require.NoError(t, err)
 
@@ -1098,7 +1096,7 @@ func createCoordinator(t testing.TB, ctx context.Context, opts ...CoordinatorOpt
 		acker = &fakeActionAcker{}
 	}
 
-	coord := New(l, nil, logp.DebugLevel, ai, specs, &fakeReExecManager{}, upgradeManager, rm, cfgMgr, varsMgr, caps, monitoringMgr, o.managed, otelComponentMgr, acker)
+	coord := New(l, nil, logp.DebugLevel, ai, specs, &fakeReExecManager{}, upgradeManager, rm, cfgMgr, varsMgr, caps, monitoringMgr, o.managed, otelMgr, acker)
 	return coord, cfgMgr, varsMgr
 }
 
@@ -1340,7 +1338,7 @@ func (f *fakeVarsManager) DefaultProvider() string {
 	return ""
 }
 
-var _ OTelComponentManager = (*fakeOtelComponentManager)(nil)
+var _ OTelManager = (*fakeOtelComponentManager)(nil)
 
 type fakeOtelComponentManager struct {
 	updateCollectorCallback func(*confmap.Conf) error
